tests: Fix store test
