+++ lib/assets/javascripts/new-dashboard/store/index.js
@@ -1,21 +1,20 @@
-import CartoNode from 'carto-node';
 import Vue from 'vue';
 import Vuex from 'vuex';
+import CartoNode from 'carto-node';
 
 // Store Modules
 import user from './user';
 import config from './config';
 import maps from './maps';
 import datasets from './datasets';
-import NotificationsFactory from './notifications/NotificationsFactory';
+import notifications from './notifications';
 
 Vue.use(Vuex);
 
-const client = new CartoNode.AuthenticatedClient();
-const notifications = NotificationsFactory.create(client);
-
 const storeOptions = {
-  state: {},
+  state: {
+    client: new CartoNode.AuthenticatedClient()
+  },
   getters: {},
   modules: {
     config,
