+++ lib/assets/test/spec/new-dashboard/unit/specs/store/helpers.js
@@ -10,6 +10,7 @@ const noop = () => {};
  * @param {Array} [expectedMutations=[]] mutations expected to be committed
  * @param {Array} [expectedActions=[]] actions expected to be dispatched
  * @param {Function} [done=noop] to be executed after the tests
+ * @param {Object} [rootstate=state] rootstate provided to the action, when not given will fallback state.
  * @return {Promise}
  *
  * @example
@@ -46,11 +47,14 @@ export const testAction = (
   state,
   expectedMutations = [],
   expectedActions = [],
-  done = noop
+  done = noop,
+  rootState
 ) => {
   if (typeof action === 'undefined') {
     throw new Error('Store Action is undefined');
   }
+
+  rootState = rootState || state; // FIXME
 
   const mutations = [];
   const actions = [];
@@ -88,7 +92,7 @@ export const testAction = (
     done();
   };
 
-  const result = action({ commit, state, dispatch, rootState: state, rootGetters: state }, payload);
+  const result = action({ commit, state, dispatch, rootState, rootGetters: state }, payload);
 
   return new Promise(resolve => {
     setImmediate(resolve);
