+++ lib/assets/test/spec/new-dashboard/unit/specs/store/notifications/notifications.spec.js
@@ -1,6 +1,7 @@
-import NotificationsFactory from 'new-dashboard/store/notifications/NotificationsFactory';
-
 import { testAction } from '../helpers';
+import notificationsStore from 'new-dashboard/store/notifications';
+
+const noop = jest.fn();
 
 describe('notificationsStore', () => {
   describe('mutations', () => {
@@ -47,15 +48,21 @@ describe('notificationsStore', () => {
     });
   });
 
-  describe('actions', () => {
+  describe.only('actions', () => {
+    let rootState;
+    beforeEach(() => {
+      rootState = {
+        client: {
+          getConfig: callback => callback(null, null, {
+            organization_notifications: ['fake_notification']
+          })
+        }
+      };
+    });
     describe('.fetchNotifications', () => {
       describe('when the request is succesful', () => {
         it('should update the state with the given notifications and mark notifications as read', () => {
-          const store = NotificationsFactory.create({
-            getConfig: callback => callback(null, null, {
-              organization_notification: ['fake_notifications']
-            })
-          });
+          const store = notificationsStore;
           const action = store.actions.fetchNotifications;
           const state = {};
           const actionPayload = null;
@@ -69,38 +76,43 @@ describe('notificationsStore', () => {
             payload: null
           }];
 
-          testAction(action, actionPayload, state, expectedMutations, expectedActions);
+          testAction(action, actionPayload, state, expectedMutations, expectedActions, noop, rootState);
         });
       });
 
       describe('when the request fails', () => {
+        const dummyError = { error: "Wrong 'order' parameter value. Valid values are one of [:updated_at, :size, :mapviews, :likes]" };
+        beforeEach(() => {
+          rootState = {
+            client: {
+              getConfig: callback => callback(dummyError)
+            }
+          };
+        });
+
         it('should update the state with the given notifications and mark notifications as read', () => {
-          const store = NotificationsFactory.create({
-            getConfig: callback => callback('fake_error') // eslint-disable-line
-          });
+          const store = notificationsStore;
           const action = store.actions.fetchNotifications;
           const state = {};
           const actionPayload = null;
           const expectedMutations = [{
             'type': 'setFetchingState'
           }, {
-            'payload': 'fake_error',
+            'payload': dummyError,
             'type': 'setRequestError'
           }];
           const expectedActions = [];
 
-          testAction(action, actionPayload, state, expectedMutations, expectedActions);
+          testAction(action, actionPayload, state, expectedMutations, expectedActions, noop, rootState);
         });
       });
     });
 
     describe('.markNotificationsAsRead', () => {
-      let clientSpy, store, action, state, actionPayload;
+      let store, action, state, actionPayload;
       beforeEach(() => {
-        clientSpy = {
-          updateNotification: jest.fn()
-        };
-        store = NotificationsFactory.create(clientSpy);
+        rootState.client.updateNotification = jest.fn();
+        store = notificationsStore;
         action = store.actions.markNotificationsAsRead;
         state = {
           notifications: [{
@@ -116,16 +128,17 @@ describe('notificationsStore', () => {
           apiKey: 'fake_api_key'
         };
       });
+
       it('should notify the server about read_at', () => {
-        testAction(action, actionPayload, state, [], []);
-        expect(clientSpy.updateNotification).toHaveBeenCalledWith('fake_user_id', 'fake_api_key', {
+        testAction(action, actionPayload, state, [], [], noop, rootState);
+        expect(rootState.client.updateNotification).toHaveBeenCalledWith('fake_user_id', 'fake_api_key', {
           'id': 'fake_id_1',
           'read_at': expect.any(String)
         });
       });
 
       it('should not modify state notifications', () => {
-        testAction(action, actionPayload, state, [], []);
+        testAction(action, actionPayload, state, [], [], noop, rootState);
         expect(state.notifications[1].read_at).toBeNull();
       });
     });
